feat: add global exception handling data violation
fix: update is now a resource not found exception

+++ src/main/java/com/bookstore/bookstore/services/ProductService.java
@@ -7,9 +7,11 @@ import com.bookstore.bookstore.services.exceptions.DatabaseException;
 import com.bookstore.bookstore.services.exceptions.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.NoSuchElementException;
@@ -53,13 +55,20 @@ public class ProductService {
             product = repository.save(product);
             return new ProductDTO(product);
         } catch (EntityNotFoundException e) {
-            throw new DatabaseException("Data Not Found");
+            throw new ResourceNotFoundException("Resource Not Found");
         }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource Not Found");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Integrity Violated");
+        }
     }
 
     private void dtoToEntity(ProductDTO dto, Product product) {
